+++ include/votca/tools/edgecontainer.h
@@ -38,7 +38,7 @@ namespace tools {
 class EdgeContainer {
  protected:
   /// The vertex, the neighboring vertices and the number of edges
-  std::unordered_map<long int, std::unordered_map<int, int>> adj_list_;
+  std::unordered_map<long int, std::unordered_map<long int, long int>> adj_list_;
 
  public:
   /// Constructors can take no arguments a single Edge or a vector of edges
@@ -53,12 +53,12 @@ class EdgeContainer {
   /// Determine the degree of the vertex/number of edges attached
   long int getDegree(const long int vertex) const;
   /// Determine if a vertex with the specified degree exists
-  bool vertexExistWithDegree(int degree) const;
+  bool vertexExistWithDegree(long int degree) const;
 
   /// Check if the edge exists returns true or false
   bool edgeExist(const Edge& edge) const;
   /// Check if the vertex exists returns true or false
-  bool vertexExist(int vertex) const;
+  bool vertexExist(long int vertex) const;
   /// Add an edge to the container
   void addEdge(Edge edge);
   /// Add a lone vertex
@@ -68,9 +68,9 @@ class EdgeContainer {
   /// Get all the vertices in vector form
   std::vector<long int> getVertices() const;
   /// Get the vertices neighboring vert
-  std::vector<int> getNeighVertices(int vertex) const;
+  std::vector<long int> getNeighVertices(long int vertex) const;
   /// Get the edges neighboring vert
-  std::vector<Edge> getNeighEdges(int vertex) const;
+  std::vector<Edge> getNeighEdges(long int vertex) const;
   /// Print output of object
   friend std::ostream& operator<<(std::ostream& os,
                                   const EdgeContainer edge_container);
